fix icon stack
