+++ packages/geoview-core/src/core/components/common/layer-icon.tsx
@@ -1,7 +1,77 @@
-import { Box, CircularProgressBase, ErrorIcon, GroupWorkOutlinedIcon } from '@/ui';
+import { Box, CircularProgressBase, ErrorIcon, GroupWorkOutlinedIcon, IconButton, BrowserNotSupportedIcon } from '@/ui';
 import { TypeLegendLayer } from '@/core/components/layers/types';
-import { IconStack } from '@/core/components/icon-stack/icon-stack';
+import { getSxClasses } from './layer-icon-style';
+import { useIconLayerSet } from '@/core/stores/store-interface-and-intial-values/layer-state';
 import { LayerListEntry } from '.';
+import { useTheme } from '@mui/material/styles';
+
+export interface TypeIconStackProps {
+  layerPath: string;
+  onIconClick?: () => void;
+  onStackIconClick?: (e: React.KeyboardEvent<HTMLElement>) => void;
+}
+
+/**
+ * Icon Stack to represent layer icons
+ *
+ * @param {string} layerPath
+ * @returns {JSX.Element} the icon stack item
+ */
+function IconStack({ layerPath, onIconClick, onStackIconClick }: TypeIconStackProps): JSX.Element | null {
+  const theme = useTheme();
+  const sxClasses = getSxClasses(theme);
+
+  const iconData = useIconLayerSet(layerPath);
+
+  const iconImage: string = iconData && iconData.length > 0 ? iconData[0] : '';
+  const iconImageStacked: string = iconData && iconData.length > 1 ? iconData[1] : '';
+  const numOfIcons: number | undefined = iconData?.length;
+
+  const iconStackContent = (): JSX.Element | null => {
+    // TODO: refactor - try to remove the nested ternary to simplify reading
+    // eslint-disable-next-line no-nested-ternary
+    return numOfIcons === 1 ? (
+      <IconButton
+        tabIndex={-1}
+        sx={sxClasses.iconPreview}
+        color="primary"
+        size="small"
+        onClick={iconImage === 'no data' ? undefined : onIconClick}
+        aria-hidden="true"
+      >
+        {iconImage === 'no data' ? (
+          <BrowserNotSupportedIcon />
+        ) : (
+          <Box sx={sxClasses.legendIcon}>
+            <img alt="icon" src={iconImage} style={sxClasses.maxIconImg} />
+          </Box>
+        )}
+      </IconButton>
+    ) : // eslint-disable-next-line no-nested-ternary
+    numOfIcons && numOfIcons > 0 ? (
+      <Box tabIndex={-1} onClick={onIconClick} sx={sxClasses.stackIconsBox} onKeyPress={(e) => onStackIconClick?.(e)} aria-hidden="true">
+        <IconButton sx={sxClasses.iconPreviewStacked} color="primary" size="small" tabIndex={-1} aria-hidden="true">
+          <Box sx={sxClasses.legendIconTransparent}>
+            {iconImageStacked && <img alt="icon" src={iconImageStacked} style={sxClasses.maxIconImg} />}
+          </Box>
+        </IconButton>
+        <IconButton sx={sxClasses.iconPreviewHoverable} color="primary" size="small" tabIndex={-1} aria-hidden="true">
+          <Box sx={sxClasses.legendIcon}>{iconImage && <img alt="icon" src={iconImage} style={sxClasses.maxIconImg} />}</Box>
+        </IconButton>
+      </Box>
+    ) : layerPath !== '' && iconData.length === 0 && layerPath.charAt(0) !== '!' ? (
+      <Box tabIndex={-1} onClick={onIconClick} sx={sxClasses.stackIconsBox} onKeyPress={(e) => onStackIconClick?.(e)} aria-hidden="true">
+        <IconButton sx={sxClasses.iconPreviewStacked} color="primary" size="small" tabIndex={-1} aria-hidden="true">
+          <Box sx={sxClasses.legendIconTransparent}>
+            <BrowserNotSupportedIcon />
+          </Box>
+        </IconButton>
+      </Box>
+    ) : null;
+  };
+
+  return iconStackContent();
+}
 
 interface LayerIconProps {
   layer: TypeLegendLayer | LayerListEntry;
