+++ packages/geoview-core/src/core/components/hover-tooltip/hover-tooltip-styles.ts
@@ -1,5 +1,7 @@
 import { Theme } from '@mui/material/styles';
 
+// ? I doubt we want to define an explicit type for style properties?
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const getSxClasses = (theme: Theme): any => ({
   tooltipItem: {
     color: theme.palette.geoViewColor.bgColor.light[900],
